requirements.txt
# cloudpickle==3.1.0
# Farama-Notifications==0.0.4
# gymnasium==1.0.0
# numpy==2.1.3
# typing_extensions==4.12.2
# pygame==2.6.1

cfgv==3.4.0
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
Farama-Notifications==0.0.4
filelock==3.13.1
fonttools==4.55.0
fsspec==2024.2.0
gymnasium==1.0.0
identify==2.6.2
Jinja2==3.1.3
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
nodeenv==1.9.1
numpy==2.1.3
packaging==24.2
pillow==10.2.0
pip==24.0
platformdirs==4.3.6
pre_commit==4.0.1
pygam==2.6.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
setuptools==70.0.0
six==1.16.0
sympy==1.13.1
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
typing_extensions==4.12.2
